api_yamdb/requirements.txt
asgiref==3.2.10
atomicwrites==1.4.1
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
gunicorn==20.0.4
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
packaging==21.3
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.11.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==0.19.0
pytz==2020.1
requests==2.26.0
sqlparse==0.3.1
toml==0.10.2
typing-extensions==4.3.0
urllib3==1.26.12
zipp==3.8.1
pytest-pythonpath